Bug fixing

+++ src/engine/mesh/mesh.c
@@ -11,8 +11,8 @@ void CFD_Setup_Mesh(CFD_t *cfd)
     cfd->engine->mesh->element->size->dx = cfd->in->geometry->x / cfd->engine->mesh->nodes->Nx;
     cfd->engine->mesh->element->size->dy = cfd->in->geometry->y / cfd->engine->mesh->nodes->Ny;
 
-    uint8_t rows = cfd->engine->mesh->nodes->Ny + 2 * cfd->engine->mesh->n_ghosts;
-    uint8_t cols = cfd->engine->mesh->nodes->Nx + 2 * cfd->engine->mesh->n_ghosts;
+    uint16_t rows = cfd->engine->mesh->nodes->Ny + 2 * cfd->engine->mesh->n_ghosts;
+    uint16_t cols = cfd->engine->mesh->nodes->Nx + 2 * cfd->engine->mesh->n_ghosts;
 
     cfd->engine->mesh->data->x = MAT_Init(
         rows,
@@ -25,12 +25,12 @@ void CFD_Setup_Mesh(CFD_t *cfd)
 
 void CFD_Generate_Mesh(CFD_t *cfd)
 {
-    uint8_t rows = cfd->engine->mesh->data->x->rows;
-    uint8_t cols = cfd->engine->mesh->data->x->cols;
+    uint16_t rows = cfd->engine->mesh->data->x->rows;
+    uint16_t cols = cfd->engine->mesh->data->x->cols;
 
-    for (uint8_t j = 0; j < rows; j++)
+    for (uint16_t j = 0; j < rows; j++)
     {
-        for (uint8_t i = 0; i < cols; i++)
+        for (uint16_t i = 0; i < cols; i++)
         {
             cfd->engine->mesh->data->x->data[j][i] = (i - cfd->engine->mesh->n_ghosts + 0.5) * cfd->engine->mesh->element->size->dx;
             cfd->engine->mesh->data->y->data[j][i] = (j - cfd->engine->mesh->n_ghosts + 0.5) * cfd->engine->mesh->element->size->dy;
