+++ src/utils/cALGEBRA/cVEC.c
@@ -13,7 +13,7 @@
 #include "cVEC.h"
 #include "../cLOG/cLOG.h"
 
-cVEC_t *VEC_Init(uint8_t length)
+cVEC_t *VEC_Init(uint16_t length)
 {
     cVEC_t *V = (cVEC_t *)malloc(sizeof(cVEC_t));
     if (V == NULL)
@@ -31,7 +31,7 @@ cVEC_t *VEC_Init(uint8_t length)
         exit(EXIT_FAILURE);
     }
 
-    for (uint8_t i = 0; i < length; i++)
+    for (uint16_t i = 0; i < length; i++)
     {
         V->data[i] = 0.0;
     }
@@ -45,7 +45,7 @@ cVEC_t *VEC_Sum(cVEC_t *V1, cVEC_t *V2)
 
     cVEC_t *result = VEC_Init(V1->length);
 
-    for (uint8_t i = 0; i < V1->length; i++)
+    for (uint16_t i = 0; i < V1->length; i++)
     {
         result->data[i] = V1->data[i] + V2->data[i];
     }
@@ -57,7 +57,7 @@ cVEC_t *VEC_Multiply(double scalar, cVEC_t *V)
 {
     cVEC_t *result = VEC_Init(V->length);
 
-    for (uint8_t i = 0; i < V->length; i++)
+    for (uint16_t i = 0; i < V->length; i++)
     {
         result->data[i] = scalar * V->data[i];
     }
@@ -69,7 +69,7 @@ cMAT_t *VEC_ToMatrix(cVEC_t *V)
 {
     cMAT_t *result = MAT_Init(V->length, 1);
 
-    for (uint8_t i = 0; i < V->length; i++)
+    for (uint16_t i = 0; i < V->length; i++)
     {
         result->data[i][0] = V->data[i];
     }
@@ -86,7 +86,7 @@ void VEC_Free(cVEC_t *V)
 void VEC_Print(cVEC_t *V)
 {
     printf("[\n");
-    for (uint8_t i = 0; i < V->length; i++)
+    for (uint16_t i = 0; i < V->length; i++)
     {
         printf("\t%f", V->data[i]);
         if (i < V->length - 1)
